Fix a variety of compiler warnings.  No logic changes, mainly just dead code and variable removal and signed/unsigned comparison issues.
git-svn-id: http://svn.openimageio.org/oiio/trunk@1254 31904243-a364-42ac-a3db-9305396518ba

+++ src/ico.imageio/icoinput.cpp
@@ -283,7 +283,7 @@ ICOInput::readimg ()
     }
 
     // otherwise we're dealing with a DIB
-    DASSERT (m_spec.scanline_bytes() == (m_spec.width * 4));
+    DASSERT (m_spec.scanline_bytes() == ((size_t)m_spec.width * 4));
     m_buf.resize (m_spec.image_bytes());
 
     //std::cerr << "[ico] DIB buffer size = " << m_buf.size () << "\n";
